selinux-policy: backstop noexec enforcement
Most mounts are mounted with the "noexec" option, to stop the host
from executing unexpected binaries. However, the mounts for container
layers cannot use this option, or else containers could not be run.

For additional protection, split up the permissions for "load" and
"execute", and ensure that only container runtimes and containers are
allowed to execute container binaries.

To make this work with overlayfs mounts, it is also necessary for the
mounts created by systemd units at startup to be mounted by a process
that has "execute" permissions. These credentials are checked before
those of any subsequent process that attempts to execute files on the
mount.

This is achieved by a transition rule for `mount` to a new domain.
Even though `mount` does not need the "execute" permission, granting
it allows the other non-container runtime processes in the host to be
restricted, which is an overall win for security.

See this upstream discussion for more on overlayfs mounter creds:
  https://lore.kernel.org/selinux/20211117015806.2192263-1-dvander@google.com/

Signed-off-by: Ben Cressey <bcressey@amazon.com>