selinux-policy: suppress audit for tmpfs relabels
Containers are not permitted to relabel files. However, they may use
standard tools such as `mv` that attempt to set xattrs on the target,
and this generates an audit message.

Since the operation is otherwise successful, this can be confusing
and suggest a problem that doesn't really exist. Turn off the audit
message to avoid this.

Signed-off-by: Ben Cressey <bcressey@amazon.com>