create mount points for container storage
Create the directories for container storage early in boot, so that
they can be mounted into bootstrap containers. This is split up by
container runtime so that e.g. `/var/lib/docker` is not created on a
host that does not include Docker.

Existing symlinks in these locations are removed. These may have been
created by bootstrap containers, and if so may point to directories
in `/mnt` that are not yet available. This would prevent bootstrap
containers from starting and mounting the intended storage device.

Signed-off-by: Ben Cressey <bcressey@amazon.com>