release: mask storage on /local
Since mounts are not propagated back to `/local`, subdirectories like
`/local/var` will no longer be identical and interchangeable with
`/var`.

This is confusing and could lead to situations where files written to
some subdirectories under `/local/var` are not visible to processes
reading from `/var`, which might look like data loss or application
bugs.

Mounting an empty, read-only directory over each directory prevents
the confusion, at the cost of potentially disrupting containers that
try to mount these directories using the `/local` prefixed paths.

That is a relatively obvious failure mode - the container will fail
to start, or encounter errors when writing to the mount - which is
preferable to subtle misbehavior with no obvious errors.

Signed-off-by: Ben Cressey <bcressey@amazon.com>