os: improve systemd reload handling
Every time a systemd unit or drop-in is rendered from a template, we
potentially need to reload systemd. However, this reload is expensive
and ideally should be avoided if necessary or done just once to cover
a batch of changes if possible.

Today we only have a single template that renders a systemd drop-in,
for the kubelet `ExecStart` command, but others could be added in the
future, and it would be easy to overlook the required systemd reload
command.

Solve both problems by adding a general mechanism that checks whether
any systemd file has been recreated, and run `systemd daemon-reload`
once before any restart commands are run.

Signed-off-by: Ben Cressey <bcressey@amazon.com>