models: add new setting to block kernel modules
For compliance or policy reasons, the system administrator might want
to prevent certain kernel modules from being loaded, if they are not
expected to be used.

This isn't meant to be an ironclad guarantee that the desired kernel
functionality will never be available. The module could be built into
the kernel instead, or loaded early in the boot, or a core dependency
of another required module.

Signed-off-by: Ben Cressey <bcressey@amazon.com>